Various improvements. Moved meals to a separate package

+++ src/edu/gcc/maplocation/MapLocation.java
@@ -19,13 +19,13 @@ public class MapLocation {
 	private long id;
 
 	private int type;
-	private int xCoord;
-	private int yCoord;
+	private double xCoord;
+	private double yCoord;
 
 	private String name;
 	private String campus;
 
-	public MapLocation(int x, int y, int type, String n, String campus) {
+	public MapLocation(double x, double y, int type, String n, String campus) {
 		this.xCoord = x;
 		this.yCoord = y;
 		this.type = type;
@@ -35,24 +35,24 @@ public class MapLocation {
 
 	// calculates the length of a vector between 2 points on Map
 	public double distance(MapLocation l) {
-		int xSum = Math.abs(xCoord - l.xCoord);
-		int ySum = Math.abs(yCoord - l.yCoord);
-		return Math.sqrt((double) (xSum * xSum) + (ySum * ySum));
+		double xSum = Math.abs(xCoord - l.xCoord);
+		double ySum = Math.abs(yCoord - l.yCoord);
+		return Math.sqrt((xSum * xSum) + (ySum * ySum));
 	}
 
-	public int getxCoord() {
+	public double getxCoord() {
 		return xCoord;
 	}
 
-	public void setxCoord(int xCoord) {
+	public void setxCoord(double xCoord) {
 		this.xCoord = xCoord;
 	}
 
-	public int getyCoord() {
+	public double getyCoord() {
 		return yCoord;
 	}
 
-	public void setyCoord(int yCoord) {
+	public void setyCoord(double yCoord) {
 		this.yCoord = yCoord;
 	}
